+++ src/main/java/module-info.java
@@ -25,4 +25,5 @@
 module comparator.extensions.main {
 	exports io.github.astrapi69.comparator;
 	exports io.github.astrapi69.comparator.factory;
+	exports io.github.astrapi69.comparator.object;
 }